requirements.txt
django-countries==7.5.1
tqdm==4.65.0
